Extract extensions into separate files
