+++ Sources/EventSpec.swift
@@ -1,0 +1,33 @@
+//
+// Copyright 2023 Adobe. All rights reserved.
+// This file is licensed to you under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may obtain a copy
+// of the License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under
+// the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+// OF ANY KIND, either express or implied. See the License for the specific language
+// governing permissions and limitations under the License.
+//
+
+
+import Foundation
+
+/// Struct defining the event specifications - contains the event type and source
+public struct EventSpec {
+    let type: String
+    let source: String
+}
+
+/// Hashable `EventSpec`, to be used as key in Dictionaries
+extension EventSpec: Hashable & Equatable {
+
+    public static func == (lhs: EventSpec, rhs: EventSpec) -> Bool {
+        return lhs.source.lowercased() == rhs.source.lowercased() && lhs.type.lowercased() == rhs.type.lowercased()
+    }
+
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(type)
+        hasher.combine(source)
+    }
+}
